feat: 질문글 신고 기능 추가. TODO:QuestionRepository count쿼리 수정

+++ src/main/java/kit/se/capstone2/reports/domain/repository/QuestionReportRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface QuestionReportRepository extends JpaRepository<QuestionReport, Long> {
 	@Query("SELECT COUNT(qr) FROM QuestionReport qr WHERE qr.question = :question")
-	int countByQuestion(Question question);
+	long countByQuestion(Question question);
 }
